PoC: exception reference cleanup
In Python 3, raised exceptions contain the associated backtrace, that
reference every objects on its stack frames. When such an exception is
bound to an object that is also referenced by the stack frame, a
reference cycle arises.

This patch attempts to clean up such reference cycles in a few different
cases.

 * strategy's exception_history
this is a list containing exceptions, where the list also appears on the
exception stack trace. To resolve the reference cycle, a new exception
is created based on the raised exception, and this new exception is
appended to the history.

 * exception bound to function-local variables
There is a pattern of error handling in the code that is prone to create
reference cycles:
  exc = None
  try:
    exception_raising_function()
  except Exception as e:
    exc = e
  if exc:
    do_some_additional_error_handling()

Here exc will already live in the function's stack frame when the
exception is raised, binding it to the exception itself creates a cycle.
Usually the do_some_additional_error_handling part can be moved right
into the exception handling section, so no need to use the additional
variable binding.

 * removing indirect exception handling (exc_info)
To make reasoning about where exception references can appear, exc_info
is replaced by passing the exception objects explicitly.